+++ pyodide-build/pyodide_build/cli/build_recipes.py
@@ -5,7 +5,10 @@ import typer  # type: ignore[import]
 
 from .. import buildall
 
+app = typer.Typer()
 
+
+@app.command()  # type: ignore[misc]
 def recipe(
     packages: list[str] = typer.Argument(
         ..., help="Packages to build, or * for all packages in recipe directory"
